Fix printf format warning
PAPPL 1.x uses the int type for num_media, PAPPL 2.x uses the size_t
type, so this code produced a warning with PAPPL 2.x. Cast to the
biggest of the two types and adapt the format string accordingly.

+++ lprint-common.c
@@ -815,7 +815,7 @@ lprintMediaUpdate(
 
   data->num_media = i;
 
-  LPRINT_DEBUG("lprintMediaUpdate: num_media=%d\n", data->num_media);
+  LPRINT_DEBUG("lprintMediaUpdate: num_media=%u\n", (unsigned)data->num_media);
   for (i = 0; i < data->num_media; i ++)
     LPRINT_DEBUG("lprintMediaUpdate: media[%d]='%s'\n", i, data->media[i]);
 }
